Refactor: deduplicate load_tests across noauto test suites
Extract the repeated load_tests function body into a shared factory
make_load_tests() in tests/_noauto_loader.py. Each noauto __init__.py
now declares only its test_packages list and calls make_load_tests().

Eliminates 5 identical 15-line blocks (noauto_cython, noauto_network,
noauto_onnx, noauto_tensorflow, noauto_torch).
