+++ tests/noauto_onnx/__init__.py
@@ -17,9 +17,8 @@ These tests are kept for manual testing and may be run in separate CI
 workflows dedicated to ONNX Runtime-based features.
 """
 
-from unittest import TestLoader, TestSuite
+from tests._noauto_loader import make_load_tests
 
-# Names of module to be tested
 test_packages: list[str] = [
     "tests.noauto_onnx.testn_spell_onnx",
     "tests.noauto_onnx.testn_tag_onnx",
@@ -27,19 +26,7 @@ test_packages: list[str] = [
     "tests.noauto_onnx.testn_transliterate_onnx",
 ]
 
-
-def load_tests(
-    loader: TestLoader, standard_tests: TestSuite, pattern: str
-) -> TestSuite:
-    """Load test protocol
-    See: https://docs.python.org/3/library/unittest.html#id1
-    """
-    suite = TestSuite()
-    for test_package in test_packages:
-        tests = loader.loadTestsFromName(test_package)
-        suite.addTests(tests)
-    return suite
-
+load_tests = make_load_tests(test_packages)
 
 if __name__ == "__main__":
     from unittest import main
